- moved `overlay-indicator` to `globals,js`
