+++ e2e/stackoverflow.spec.js
@@ -1,6 +1,5 @@
 import { Expect, test } from '../fixtures.js';
-import { advisoriesNames } from '../src/globals.js';
-import { tooltipSourceSelector } from './selectors.js';
+import { advisoriesNames, tooltipSourceSelector } from '../src/globals.js';
 
 const stackOverflowLink = 'https://stackoverflow.com/questions/9023672/how-do-i-resolve-cannot-find-module-error-using-node-js';
